Refactor few widgets to dedicated modules
WrappedLabel, TooltipLabel, HelpButton, ConfigureButton, BannerPanel
