+++ library/lua/gui/widgets.lua
@@ -17,6 +17,11 @@ EditField = require('gui.widgets.edit_field')
 HotkeyLabel = require('gui.widgets.hotkey_label')
 Label = require('gui.widgets.label')
 Scrollbar = require('gui.widgets.scrollbar')
+WrappedLabel = require('gui.widgets.wrapped_label')
+TooltipLabel = require('gui.widgets.tooltip_label')
+HelpButton = require('gui.widgets.help_button')
+ConfigureButton = require('gui.widgets.configure_button')
+BannerPanel = require('gui.widgets.banner_panel')
 
 makeButtonLabelText = Label.makeButtonLabelText
 parse_label_text = Label.parse_label_text
@@ -32,185 +37,6 @@ local function map_opttab(tab,idx)
         return tab[idx]
     else
         return idx
-    end
-end
-
-------------------
--- WrappedLabel --
-------------------
-
----@class widgets.WrappedLabel.attrs: widgets.Label.attrs
----@field text_to_wrap? string|string[]|fun(): string|string[]
----@field indent integer
-
----@class widgets.WrappedLabel.attrs.partial: widgets.WrappedLabel.attrs
-
----@class widgets.WrappedLabel: widgets.Label, widgets.WrappedLabel.attrs
----@field super widgets.Label
----@field ATTRS widgets.WrappedLabel.attrs|fun(attributes: widgets.WrappedLabel.attrs.partial)
----@overload fun(init_table: widgets.WrappedLabel.attrs.partial): self
-WrappedLabel = defclass(WrappedLabel, Label)
-
-WrappedLabel.ATTRS{
-    text_to_wrap=DEFAULT_NIL,
-    indent=0,
-}
-
-function WrappedLabel:getWrappedText(width)
-    -- 0 width can happen if the parent has 0 width
-    if not self.text_to_wrap or width <= 0 then return nil end
-    local text_to_wrap = getval(self.text_to_wrap)
-    if type(text_to_wrap) == 'table' then
-        text_to_wrap = table.concat(text_to_wrap, NEWLINE)
-    end
-    return text_to_wrap:wrap(width - self.indent, {return_as_table=true})
-end
-
-function WrappedLabel:preUpdateLayout()
-    self.saved_start_line_num = self.start_line_num
-end
-
--- we can't set the text in init() since we may not yet have a frame that we
--- can get wrapping bounds from.
-function WrappedLabel:postComputeFrame()
-    local wrapped_text = self:getWrappedText(self.frame_body.width-3)
-    if not wrapped_text then return end
-    local text = {}
-    for _,line in ipairs(wrapped_text) do
-        table.insert(text, {gap=self.indent, text=line})
-        -- a trailing newline will get ignored so we don't have to manually trim
-        table.insert(text, NEWLINE)
-    end
-    self:setText(text)
-    self:scroll(self.saved_start_line_num - 1)
-end
-
-------------------
--- TooltipLabel --
-------------------
-
----@class widgets.TooltipLabel.attrs: widgets.WrappedLabel.attrs
----@field show_tooltip? boolean|fun(): boolean
-
----@class widgets.TooltipLabel.attrs.partial: widgets.TooltipLabel.attrs
-
----@class widgets.TooltipLabel: widgets.WrappedLabel, widgets.TooltipLabel.attrs
----@field super widgets.WrappedLabel
----@field ATTRS widgets.TooltipLabel.attrs|fun(attributes: widgets.TooltipLabel.attrs.partial)
----@overload fun(init_table: widgets.TooltipLabel.attrs.partial): self
-TooltipLabel = defclass(TooltipLabel, WrappedLabel)
-
-TooltipLabel.ATTRS{
-    show_tooltip=DEFAULT_NIL,
-    indent=2,
-    text_pen=COLOR_GREY,
-}
-
-function TooltipLabel:init()
-    self.visible = self.show_tooltip
-end
-
-----------------
--- HelpButton --
-----------------
-
----@class widgets.HelpButton.attrs: widgets.Panel.attrs
----@field command? string
-
----@class widgets.HelpButton.attrs.partial: widgets.HelpButton.attrs
-
----@class widgets.HelpButton: widgets.Panel, widgets.HelpButton.attrs
----@field super widgets.Panel
----@field ATTRS widgets.HelpButton.attrs|fun(attributes: widgets.HelpButton.attrs.partial)
----@overload fun(init_table: widgets.HelpButton.attrs.partial): self
-HelpButton = defclass(HelpButton, Panel)
-
-HelpButton.ATTRS{
-    frame={t=0, r=1, w=3, h=1},
-    command=DEFAULT_NIL,
-}
-
-local button_pen_left = to_pen{fg=COLOR_CYAN,
-    tile=curry(textures.tp_control_panel, 7) or nil, ch=string.byte('[')}
-local button_pen_right = to_pen{fg=COLOR_CYAN,
-    tile=curry(textures.tp_control_panel, 8) or nil, ch=string.byte(']')}
-local help_pen_center = to_pen{
-    tile=curry(textures.tp_control_panel, 9) or nil, ch=string.byte('?')}
-local configure_pen_center = dfhack.pen.parse{
-    tile=curry(textures.tp_control_panel, 10) or nil, ch=15} -- gear/masterwork symbol
-
-function HelpButton:init()
-    self.frame.w = self.frame.w or 3
-    self.frame.h = self.frame.h or 1
-
-    local command = self.command .. ' '
-
-    self:addviews{
-        Label{
-            frame={t=0, l=0, w=3, h=1},
-            text={
-                {tile=button_pen_left},
-                {tile=help_pen_center},
-                {tile=button_pen_right},
-            },
-            on_click=function() dfhack.run_command('gui/launcher', command) end,
-        },
-    }
-end
-
----------------------
--- ConfigureButton --
----------------------
-
----@class widgets.ConfigureButton.attrs: widgets.Panel.attrs
----@field on_click? function
-
----@class widgets.ConfigureButton.attrs.partial: widgets.ConfigureButton.attrs
-
----@class widgets.ConfigureButton: widgets.Panel, widgets.ConfigureButton.attrs
----@field super widgets.Panel
----@field ATTRS widgets.ConfigureButton.attrs|fun(attributes: widgets.ConfigureButton.attrs.partial)
----@overload fun(init_table: widgets.ConfigureButton.attrs.partial): self
-ConfigureButton = defclass(ConfigureButton, Panel)
-
-ConfigureButton.ATTRS{
-    on_click=DEFAULT_NIL,
-}
-
-function ConfigureButton:preinit(init_table)
-    init_table.frame = init_table.frame or {}
-    init_table.frame.h = init_table.frame.h or 1
-    init_table.frame.w = init_table.frame.w or 3
-end
-
-function ConfigureButton:init()
-    self:addviews{
-        Label{
-            frame={t=0, l=0, w=3, h=1},
-            text={
-                {tile=button_pen_left},
-                {tile=configure_pen_center},
-                {tile=button_pen_right},
-            },
-            on_click=self.on_click,
-        },
-    }
-end
-
------------------
--- BannerPanel --
------------------
-
----@class widgets.BannerPanel: widgets.Panel
----@field super widgets.Panel
-BannerPanel = defclass(BannerPanel, Panel)
-
----@param dc gui.Painter
-function BannerPanel:onRenderBody(dc)
-    dc:pen(COLOR_RED)
-    for y=0,self.frame_rect.height-1 do
-        dc:seek(0, y):char('[')
-        dc:seek(self.frame_rect.width-1):char(']')
     end
 end
